stocks details

+++ app/api/transaction_routes.py
@@ -71,9 +71,9 @@ def sell_shares(id): #need to add id back later
     form['csrf_token'].data = request.cookies['csrf_token']
 
     if form.validate_on_submit():
-        num_shares = form.data['num_shares']
+        num_shares = Decimal(form.data['num_shares'])
         price_at_transaction = Stock.query.get(id).stock_to_dict()['i_price']
-        total_price = float(price_at_transaction) * num_shares
+        total_price = Decimal(price_at_transaction) * num_shares
 
         if asset.num_shares >= num_shares:
             wallet.amount += Decimal(total_price)
@@ -111,7 +111,7 @@ def cash_out(id): #will add id back later
     asset = Asset.query.filter(Asset.stock_id == id, Asset.user_id == user_id).one()
     # num_shares = asset['num_shares']
     price_at_transaction = Stock.query.get(id).stock_to_dict()['i_price']
-    total_price = float(price_at_transaction) * float(asset.num_shares)
+    total_price = Decimal(price_at_transaction) * Decimal(asset.num_shares)
 
     wallet.amount += Decimal(total_price)
